LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flexbox-min-height-auto-004-expected.txt


PASS overflow-x:visible: overflow-y stays visible; min-height:auto holds; no scrollbar gutter
FAIL overflow-x:hidden: forces overflow-y to auto; min-height:auto resolves to 0; width includes scrollbar gutter assert_equals: offsetWidth expected 59 but got 44
FAIL overflow-x:scroll: forces overflow-y to auto; min-height:auto resolves to 0; width includes scrollbar gutter assert_equals: offsetWidth expected 59 but got 44
FAIL overflow-x:auto: forces overflow-y to auto; min-height:auto resolves to 0; width includes scrollbar gutter assert_equals: offsetWidth expected 59 but got 44
